requirements.txt
Flask
langchain-community
langchain-openai
pypdf
python-dotenv
pymupdf
faiss-cpu
